+++ lib/pavilion/commands/result.py
@@ -7,7 +7,9 @@ import io
 import pathlib
 import pprint
 import shutil
-from typing import List, IO
+from math import log10, floor
+import re
+from typing import List, IO, Union, Optional
 
 from pavilion.errors import TestConfigError, ResultError
 from pavilion import cmd_utils
@@ -22,6 +24,46 @@ from pavilion import utils
 from pavilion.status_file import STATES
 from pavilion.test_run import TestRun
 from .base_classes import Command
+
+
+def _num_digits(n: int) -> int: # pylint: disable=invalid-name
+    """Only attempt to count number digits for ints, since
+    precision error makes it nonsensical for floats."""
+    if n < 0:
+        n = abs(n)
+    elif n == 0:
+        return 1
+    return floor(log10(n)) + 1
+
+
+def format_numeric(value: Union[int, float], digits: Optional[int]) -> str:
+    if digits is None:
+        return str(value)
+
+    sci_fmt = '{' + f':.{digits-1}e' + '}'
+
+    if isinstance(value, int):
+        if _num_digits(value) > digits:
+            res = sci_fmt.format(value)
+        else:
+            res = str(value)
+    elif isinstance(value, float):
+        if abs(value) < 10**-(digits - 1):
+            res = sci_fmt.format(value)
+        else:
+            res = str(round(value, digits - 1))
+    else:
+        raise ValueError(f'Cannot format value {value}. Expected one of (int, float),' +
+            f' but received {type(value)}.')
+
+    # Remove leading zeros in exponent, as well as + sign, if present
+    pos_regex = 'e\+0*'
+    neg_regex = 'e-0*'
+
+    res = re.sub(pos_regex, "e", res)
+    res = re.sub(neg_regex, "e-", res)
+
+    return res
 
 
 class ResultsCommand(Command):
@@ -101,11 +143,11 @@ class ResultsCommand(Command):
                  "filter argument will override that."
         )
         parser.add_argument(
-            '-d', '--num-digits', dest="num_digits",
+            '-d', '--max-digits', dest="max_digits", default=5,
             help="The maximum number of digits to display when formatting floating point values. If"
                  " set, numbers much larger or much smaller than 1 are displayed in scientific "
-                 "notation with the specified precision, and numbers near 1 are truncated. If not "
-                 "provided, defaults to standard Python rules for formatting numeric values."""
+                 "notation with the specified precision, and numbers near 1 are truncated. "
+                 "Defaults to 5 digits."""
         )
         filters.add_test_filter_args(parser)
 
@@ -235,7 +277,7 @@ class ResultsCommand(Command):
                 fields=fields,
                 rows=flat_sorted_results,
                 title=title_str,
-                num_digits=args.num_digits
+                default_format=lambda x: format_numeric(x, args.max_digits)
             )
 
         if args.show_log:
